refactor: recoil 상태 전체 정리
