+++ src/states/cartProducts/selector.ts
@@ -1,16 +1,7 @@
 import { selector, selectorFamily } from 'recoil';
 
-import productApis from '../../../apis/products';
-import { findTargetProduct } from '../util';
-import { cartProductState } from '../atoms';
-
-export const productSelector = selector({
-  key: 'productSelector',
-  get: async () => {
-    const data = await productApis.get();
-    return data;
-  },
-});
+import { findTargetProduct } from './util';
+import { cartProductState } from './atom';
 
 export const cartProductCountState = selector({
   key: 'cartProductCountState',
