+++ src/states/products/index.ts
@@ -1,2 +1,1 @@
-export * from './atom';
 export * from './selector';
